update: navbar design update and create subNavbar
