+++ src/layout/Root.jsx
@@ -3,12 +3,10 @@ import Navbar from "../pages/shared/Navbar";
 import Footer from "../pages/shared/Footer";
 import { useState, useEffect } from "react";
 import { FaArrowUp } from "react-icons/fa";
+import SubNavbar from "../pages/shared/SubNavbar";
 
 function Root() {
-
-  
   const [isVisible, setIsVisible] = useState(false);
-
 
   useEffect(() => {
     const toggleVisibility = () => {
@@ -23,7 +21,6 @@ function Root() {
     return () => window.removeEventListener("scroll", toggleVisibility);
   }, []);
 
-
   const scrollToTop = () => {
     window.scrollTo({
       top: 0,
@@ -33,15 +30,16 @@ function Root() {
 
   return (
     <div className="dark:bg-darkBlue relative">
+      <SubNavbar />
       <Navbar />
       <Outlet />
       <Footer />
-      
 
       {isVisible && (
         <button
           onClick={scrollToTop}
-          className="fixed xl:flex lg:flex md:flex sm:hidden hidden bottom-8 right-8 bg-blue text-white p-3 rounded-full hover:bg-blue-700 transition-all duration-300 hover:-translate-y-1">
+          className="fixed xl:flex lg:flex md:flex sm:hidden hidden bottom-8 right-8 bg-blue text-white p-3 rounded-full hover:bg-blue-700 transition-all duration-300 hover:-translate-y-1"
+        >
           <FaArrowUp />
         </button>
       )}
